+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### NEXT_VERSION_DESCRIPTION_BEGIN
 * `ScheduledTaskLifecycleListener` introduced - that helps to observe task execution, eg: configuring tracing, logging,
   monitoring, etc.
+* Tracing support implemented in a separate gradle module - 'db-queue-scheduler-brave'.
 ### NEXT_VERSION_DESCRIPTION_END
 ## [0.2.0](https://github.com/yoomoney/db-queue-scheduler/pull/2) (27-10-2021)
 
@@ -10,7 +11,9 @@ configuration entry-point;
 * Implemented core library API: `Scheduler`, `ScheduledTask`, `ScheduledTaskIdentity` and `ScheduledTaskExecutionResult`;
 * Implemented core library functionality - `db-queue` library integration, scheduling tasks feature;
 * `CronNextExecutionTimeProvider` implemented
-
+* `ScheduledTaskLifecycleListener` introduced - that helps to observe task execution, eg: configuring tracing, logging,
+monitoring, etc.
+### NEXT_VERSION_DESCRIPTION_END
 ## [0.1.0](https://github.com/yoomoney/db-queue-scheduler/pull/1) (26-10-2021)
 
 * Library skeleton set up: gradle and library-project-plugin, travis, git attributes, license.